Replace job queuing algorithm
Each job has to:
* read the dictionary
* parse patterns
* parse forbidden

The original algorithm would run n jobs concurrently.
Each job would handle 8 files at a time.

One reason for splitting the work was because I was afraid
that one runner would be stuck with a bunch of huge files,
and the other runners would finish leaving the world waiting.

It turns out that the initialization cost is a huge portion of
the work.

For repositories with *lots* of files, it dwarfs the cost of
actually checking files.

The new algorithm instead aims to use 2 queues per concurrent
job. If one early job gets stuck with lots of work, the other
jobs will end up getting the work that was intended for the
overworked worker.

The new algorithm still can lose if the heavy files are at the
end of the last work hunk.

Going forward, I might reorder the queue to put the largest
files at the front of the queues. This is probably doable now
as there is an early pass that thinks about each file without
opening it, which means it can check the file size and use
that to help build the queue order.

I'm not sure that additional change will ever be implemented.
The time improvements from this commit are so good that I'm
skeptical that the extra work to improve the algorithm will
be worth implementing.